ECA-2333: Massive fix for a seemingly small issue.
AccessUserAspectData has now gained the column tokenType, which denotes the type of token that that particular aspect interfaces to. AccessUserAspectData is now also agnostic of which AccessMatchType it represents, and merely replies with tokenType and the integer value for the match type when asked (more on that later). 

A check for this type has been added to AccessTreeNode. The check itself is performed in the relevant AuthenticationToken, meaning that AlwaysAllowAuthenticationToken can still answer true for any match requests. This is tested in AccessControlSessionBeanTest

An interface (AccessMatchValue) has been extracted from X500PrincipalMatchValue, and the implementation CliUserAccessMatchValue has been added to correspond to CliAuthenticationToken. The default aspect created in the database ("Super Admin" and "ejbca") have been designated as x500 and CLI-users respectively in the table AccessUserAspectData.

In order to be able to perform reverse lookup to get an actual match type from a known token type and integer (see above), the singleton AccessMatchValueReverseLookupRegistry has been created, which uses reflection to take a callback from a method to translate an integer to a match type. Registry still lacks a good way of detecting existing match values: this is still an outstanding issue for this ticket. Currently it contains some explicit declarations an as a result an unwanted dependency. The registry has been thoroughly unit tested. 

All ORM files and upgrade scripts have been updated, hopefully correctly.




git-svn-id: https://svn.cesecore.eu/svn/ejbca/trunk/ejbca@13034 37aeb1de-b14b-0410-9a0e-82b28170d228